The default array_filter was changed to do what it is meant to do
The default array_filter removes everything considered NULL when I only wanted it to remove actual NULL's so added a manual call back which only checks if a value is null. If not then it will return the correct string. Noticed array_filter with the default callback has been used in numerous places so may be worth checking them out